+++ src/main/java/com/raja/lib/acc/repository/LedgerRepository.java
@@ -3,5 +3,5 @@ package com.raja.lib.acc.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.raja.lib.acc.model.Ledger;
 
-public interface LedgerRepository extends JpaRepository<Ledger, Long> {
+public interface LedgerRepository extends JpaRepository<Ledger, Integer> {
 }
